+++ src/SqlArtisan/Internal/SqlBuilder/DbmsDialect/IDbmsDialect.cs
@@ -1,0 +1,6 @@
+namespace SqlArtisan;
+
+internal interface IDbmsDialect
+{
+    string GetParameterName(int index);
+}