+++ src/SqlArtisan/Internal/SqlBuilder/DbmsDialect/PostgreSqlDialect.cs
@@ -1,0 +1,6 @@
+namespace SqlArtisan;
+
+internal sealed class PostgreSqlDialect : IDbmsDialect
+{
+    public string GetParameterName(int index) => $":{index}";
+}