+++ src/SqlArtisan/Internal/SqlBuilder/DbmsDialect/SqlServerDialect.cs
@@ -1,0 +1,6 @@
+namespace SqlArtisan;
+
+internal sealed class SqlServerDialect : IDbmsDialect
+{
+    public string GetParameterName(int index) => $"@{index}";
+}