+++ src/SqlArtisan/Internal/SqlBuilder/DbmsDialect/SqliteDialect.cs
@@ -1,0 +1,6 @@
+namespace SqlArtisan;
+
+internal sealed class SqliteDialect : IDbmsDialect
+{
+    public string GetParameterName(int index) => $":{index}";
+}