+++ src/SqlArtisan/Internal/SqlBuilder/SqlBuilderBase.cs
@@ -1,17 +1,15 @@
-﻿namespace SqlArtisan.Internal;
+namespace SqlArtisan.Internal;
 
 internal abstract class SqlBuilderBase(SqlPart part)
 {
     private readonly List<SqlPart> _parts = [part];
 
-    protected void AddPart(SqlPart part)
-    {
-        _parts.Add(part);
-    }
+    protected void AddPart(SqlPart part) => _parts.Add(part);
 
-    protected SqlStatement BuildCore()
+    protected SqlStatement BuildCore(Dbms dbms)
     {
-        using SqlBuildingBuffer buffer = new();
+        IDbmsDialect dialect = DbmsDialectFactory.Create(dbms);
+        using SqlBuildingBuffer buffer = new(dialect);
         return buffer
             .AppendSpaceSeparated(_parts)
             .ToSqlStatement();