+++ src/SqlArtisan/SqlBuilder/Dbms.cs
@@ -1,0 +1,11 @@
+﻿namespace SqlArtisan;
+
+public enum Dbms
+{
+    Unknown,
+    MySql,
+    Oracle,
+    PostgreSql,
+    Sqlite,
+    SqlServer,
+}