+++ src/SqlArtisan/SqlBuilder/ISqlBuilder.cs
@@ -1,6 +1,6 @@
-﻿namespace SqlArtisan;
+namespace SqlArtisan;
 
 public interface ISqlBuilder
 {
-    SqlStatement Build();
+    SqlStatement Build(Dbms dbms = Dbms.PostgreSql);
 }